+++ tests/api-resources/orders.test.ts
@@ -8,7 +8,7 @@ const client = new ParsecAPI({
 });
 
 describe('resource orders', () => {
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('create: only required params', async () => {
     const responsePromise = client.orders.create({
       exchange: 'exchange',
@@ -27,7 +27,7 @@ describe('resource orders', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('create: required and optional params', async () => {
     const response = await client.orders.create({
       exchange: 'exchange',
@@ -36,11 +36,19 @@ describe('resource orders', () => {
       price: 0,
       side: 'buy',
       size: 0,
+      credentials: {
+        api_key_id: 'api_key_id',
+        clob_api_key: 'clob_api_key',
+        clob_api_passphrase: 'clob_api_passphrase',
+        clob_api_secret: 'clob_api_secret',
+        private_key: 'private_key',
+      },
       params: { foo: 'string' },
+      'X-Exchange-Credentials': 'X-Exchange-Credentials',
     });
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('retrieve: only required params', async () => {
     const responsePromise = client.orders.retrieve('order_id', { exchange: 'exchange' });
     const rawResponse = await responsePromise.asResponse();
@@ -52,12 +60,15 @@ describe('resource orders', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('retrieve: required and optional params', async () => {
-    const response = await client.orders.retrieve('order_id', { exchange: 'exchange' });
+    const response = await client.orders.retrieve('order_id', {
+      exchange: 'exchange',
+      'X-Exchange-Credentials': 'X-Exchange-Credentials',
+    });
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('list: only required params', async () => {
     const responsePromise = client.orders.list({ exchange: 'exchange' });
     const rawResponse = await responsePromise.asResponse();
@@ -69,12 +80,16 @@ describe('resource orders', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('list: required and optional params', async () => {
-    const response = await client.orders.list({ exchange: 'exchange', market_id: 'market_id' });
+    const response = await client.orders.list({
+      exchange: 'exchange',
+      market_id: 'market_id',
+      'X-Exchange-Credentials': 'X-Exchange-Credentials',
+    });
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('cancel: only required params', async () => {
     const responsePromise = client.orders.cancel('order_id', { exchange: 'exchange' });
     const rawResponse = await responsePromise.asResponse();
@@ -86,8 +101,11 @@ describe('resource orders', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('cancel: required and optional params', async () => {
-    const response = await client.orders.cancel('order_id', { exchange: 'exchange' });
+    const response = await client.orders.cancel('order_id', {
+      exchange: 'exchange',
+      'X-Exchange-Credentials': 'X-Exchange-Credentials',
+    });
   });
 });