+++ tests/api-resources/trades.test.ts
@@ -8,7 +8,7 @@ const client = new ParsecAPI({
 });
 
 describe('resource trades', () => {
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('list: only required params', async () => {
     const responsePromise = client.trades.list({ parsec_id: 'parsec_id' });
     const rawResponse = await responsePromise.asResponse();
@@ -20,10 +20,11 @@ describe('resource trades', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('list: required and optional params', async () => {
     const response = await client.trades.list({
       parsec_id: 'parsec_id',
+      cursor: 'cursor',
       end_ts: 0,
       limit: 1,
       outcome: 'outcome',